DNA Streamline Update
Fixed so that all the files don't get removed

This update alters the way the HomeScreen and Common files work
on the CrossWalk platform while retaining basic functionality.
Most components have been removed. Topbar and Bottombar remain.
The launchApplication and getAppByName functions still exist.

Items to check:
The DNA element should be controllable with touchscreen controls.
The Logo button should end the current app on a long touch. (2.5s)
The HomeScreen cannot be closed by the user.
The Back button should return to the homescreen on touch.
The Back button should appear as the Tizen Logo while on the home screen.
The App Grid button should show the App Grid on touch.
The App Grid should launch applications on touch.
The Settings button should show the Settings Icon Panel on touch.
The Volume Indicator should show the Volume Slider on touch.
The Volume Slider should respond to touch events.
The TopBar should show the first 7 installed applications and touch to launch.
The Volume and Settings options will need to be reattached.

The ID for the HomeScreen is now JLRPOCX001.HomeScreen

Change-Id: I86bfde9848f9e5d25721ad36ea7705d750d66a69