Weather App Fixes
This patch fixes a few of the problems the Weather App has been having.

New in this patch:

The Weather Bands come up centered when they are added and when the app is started.

The Minus Button moves all bands to the center-display where the X-buttons are.
When toggling off the X-buttons, all bands are not moved to center.

The Plus Button toggles the X-buttons off before allowing the user to add.
This prevents the X-buttons from getting out-of-sync.

Touch-events now fire for the delete-location buttons!

Change-Id: Ic2768906ae7f9a51a3a0363f8a40c055c1109cb2