Ember JavaScript Streamline
The scripts need to be loaded before the document ready event fires, so they
cannot be moved out of the index document. There may be a way to load them
in a serialized fashion by using a Require library, but that may just add
unnecessary obfuscation for what is otherwise a superficial improvement.

That being said, a number of libraries were removed (via commenting) from
AppViews.js, and those javascript files have been removed from the index
document and filesystem herein.

Change-Id: Ie264ffea84d4317139e014ac1e2d80ae480349c8