TopBar Drag and Drop
This update allows application icons to be added to the topBar from the App Grid.
The topBar starts in an empty state, and can be updated by pressing an icon on the App Grid
and dragging it up to the topBar.

Known issues with this implementation:
LocalSettings issue:
Icon settings do not persist between applications or installs.
TouchPunch issue:
Touch events no longer launch applications but mouse events work properly.
Implementation issue:
Icons overwrite the icons that were in the spot before, instead of pushing them aside.

Change-Id: I832851b18fe80318aa20c963dfc3dac531c6d71a