TopBar Looks Nice but still doesn't click
This update fixes the topbar and app grid layout issues. The app grid also
allows the user to touch icons in order to open them. Icons can additionally
be dragged from the app grid into the topbar by touching an icon firmly for
one second, and without letting go, dragging that chosen icon up to the empty
topbar area.

Known issue: Icons in the topbar can be rearranged, but cannot be 'clicked'.

Change-Id: Ie579496f0c838f0c86e26ea32b9604e33ac1b593