Topbar Drag and Drop without Persistence
All issues with Dragging and Dropping clickable icons are resolved.
Now, the app grid, which looks great, features icons that can be clicked, but also,
dragged to the topbar, which looks great, and once those icons are on the topbar,
they can also be rearranged, or removed from the topbar by dragging them off of it.

The icons do not persist between apps, or between executions of the same app.

Change-Id: I737c001482a4c2a2b7526af5dddd9462c0100872