Settings Fixes
This patch updates the settings panel so that the App Grid closes when the Settings icon is clicked.
Also, the volume indicator closes when the app grid is fixed.
Also, when the Settings panel closes and is re-opened, the list of settings is shown instead of whatever panel was open last.

Change-Id: I96ec8f92e97fb6344a2856349c786bb2242530e9