Topbar and Bottombar updates
This patch adds a Drag and Drop topbar, and easily accessible volume functions for MOST.

Known issues:
Icon positions are not shared between Running applications.
This is due to the "Configuration" storage method being out of sync, not a JavaScript issue in topBar.
The issue resolves itself when the applications are closed and reopened.

Change-Id: I59e7e3d9b4cbee1879e4ad3b0bfa98621728399d